cron: Take screenshots of all displays daily at 3:00 PM.
If run `~/.local/bin/screenshot_all_displays.sh` directly in crontab,
it can only capture the desktop wallpaper, not the apps that are
running. This is because crontab does not have the permission to record
the screen.

Therefore, [Automator][1] was used to solve this problem. It
can run shell scripts, and after saving it as an application, it can
use `open` to call and run it. The first time it runs, it will ask you
for permission as an application. This way, it can take screenshots in
crontab.

[1]: https://en.wikipedia.org/wiki/Automator_(macOS)